move session file to services directory & marked services
