+++ src/services/session.go
@@ -1,4 +1,4 @@
-package database
+package services
 
 import (
 	"github.com/gofiber/fiber/v2"
